% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{cameraPR_cmp}
\alias{cameraPR_cmp}
\title{Run cameraPR for one `comparison`}
\usage{
cameraPR_cmp(v, sets, comparison, lfc)
}
\arguments{
\item{sets}{a named list of character vectors, each specifying a set of gene
names (ie., a signature)}

\item{comparison}{a character vector, see [comp2contrast()]}

\item{lfc}{LFC threshold, passed to [limma::treat()]}

\item{voom}{object (EList)}
}
\description{
Run cameraPR for one `comparison`
}
