% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{comp2contrast}
\alias{comp2contrast}
\title{Convert a comparison definition to a contrast for use with limma and camera}
\usage{
comp2contrast(comparison, design)
}
\arguments{
\item{comparison}{A comparison definition is a list of character vectors.
     names(comparisons) defines a label for the comparison and should include
     "-vs-", for example: "KO-vs-WT".

     comparison[1] : the variable name defining the groups
     comparison[2] : the level for the first group
     comparison[3] : the level for the second group
     For example, a design ~ 0 + group could work with:

     comparison <- c("group", "KO", "WT")

     to compare "KO" to "WT", which are both levels in "group".}

\item{design}{formula object of the underlying design}
}
\description{
Convert a comparison definition to a contrast for use with limma and camera
}
\seealso{
[limma::makeContrasts()]
}
