% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_pdf.R
\name{fig_pdf}
\alias{fig_pdf}
\title{Write a plot to a PDF device based on the current knitr chunk options}
\usage{
fig_pdf(outdir, p, name = NULL, suffix = NULL)
}
\arguments{
\item{outdir}{Output directory}
}
\description{
Write a plot to a PDF device based on the current knitr chunk options
}
