% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{plt_barcode}
\alias{plt_barcode}
\title{Plot camera results with all genes for signatures as a "barcode plot",
showing the underlying statistic on which the enrichment score is based.}
\usage{
plt_barcode(dat, comparisons = NULL, sets = NULL, dat_camera = NULL)
}
\arguments{
\item{comparisons}{*Names* of comparisons that were passed to [run_camera()]. These are
plotted, other comparisons are omitted. If NULL, plot all.}

\item{sets}{a named list of character vectors, each specifying a set of gene
names (ie., a signature)}

\item{dat_camera}{results from [run_camera()]}
}
\description{
Plot camera results with all genes for signatures as a "barcode plot",
showing the underlying statistic on which the enrichment score is based.
}
