% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{plt_camera_hm}
\alias{plt_camera_hm}
\title{Plot camera results as a heatmap.}
\usage{
plt_camera_hm(dat_camera, comparisons = NULL, sets = NULL, col_lim = 10)
}
\arguments{
\item{dat_camera}{results from [run_camera()]}

\item{comparisons}{*Names* of comparisons that were passed to [run_camera()]. These are
plotted, other comparisons are omitted. If NULL, plot all.}

\item{sets}{a named list of character vectors, each specifying a set of gene
names (ie., a signature)}

\item{col_lim}{limit for the color scale (absolute value, applies to
positive and negative)}
}
\description{
Plot camera results as a heatmap.
}
