% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{plt_camera_overall}
\alias{plt_camera_overall}
\title{Overall enrichment of signatures in `comparisons`}
\usage{
plt_camera_overall(dat_camera, comparisons = NULL, alpha)
}
\arguments{
\item{dat_camera}{results from [run_camera()]}

\item{comparisons}{*Names* of comparisons that were passed to [run_camera()]. These are
plotted, other comparisons are omitted. If NULL, plot all.}

\item{alpha}{FDR threshold at which to draw a line in the plot}
}
\description{
Overall enrichment of signatures in `comparisons`
}
