% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{plt_camera_top}
\alias{plt_camera_top}
\title{Plot most significantly enriched signatures in either direction.}
\usage{
plt_camera_top(dat_camera, comparisons = NULL, alpha, ntop)
}
\arguments{
\item{dat_camera}{results from [run_camera()]}

\item{comparisons}{*Names* of comparisons that were passed to [run_camera()]. These are
plotted, other comparisons are omitted. If NULL, plot all.}

\item{alpha}{FDR threshold at which to draw a line in the plot}

\item{ntop}{Maximum number of top signatures to show. All must meet `alpha`.}
}
\description{
Plot most significantly enriched signatures in either direction.
}
