% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pval_fmt}
\alias{pval_fmt}
\title{Transform scientific notation (1e-3) to an exponential expression (10^-3)}
\usage{
pval_fmt(x, thresh = 1e-04, exact = FALSE)
}
\description{
Transform scientific notation (1e-3) to an exponential expression (10^-3)
}
