% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{run_camera}
\alias{run_camera}
\title{Run cameraPR for all `comparisons`}
\usage{
run_camera(v, markers, comparisons, lfc)
}
\arguments{
\item{lfc}{LFC threshold, passed to [limma::treat()]}
}
\description{
Run cameraPR for all `comparisons`
}
