% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_signatures.R
\name{run_limma}
\alias{run_limma}
\title{Run limma for all `comparisons`}
\usage{
run_limma(v, comparisons, lfc)
}
\arguments{
\item{comparisons}{a list of character vectors, see [comp2contrast()]}

\item{lfc}{LFC threshold, passed to [limma::treat()]}
}
\description{
Run limma for all `comparisons`
}
