% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatSummaryPanel.R
\name{stat_summary_panel}
\alias{stat_summary_panel}
\title{A stat layer for ggplot2 that allows calculation of summary statistics per
panel, with access to the whole data.
export}
\usage{
stat_summary_panel(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
A stat layer for ggplot2 that allows calculation of summary statistics per
panel, with access to the whole data.
export
}
